backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.49
alembic==1.13.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
fastapi-mail==1.4.1
python-decouple==3.8
asyncpg==0.29.0
databases[postgresql]==0.8.0
aiofiles==23.2.1
Pillow==10.1.0 